pyproject.toml
[tool.poetry]
name = "ModePlait"
version = "0.1.0"
description = ""
authors = ["naoki <74347818+C-naoki@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[[tool.poetry.source]]
name = "torch_cu117"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"

[project]
name = "ModePlait"
version = "0.1.0"
description = "Personal Python Template"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.15"
pip = "==23.2.1"
numpy = "1.23.5"
pandas = "1.5.3"  # for pytorch-forecasting
matplotlib = "3.8.2"
scikit-learn = "1.1.3"  # for pytorch-forecasting
seaborn = "0.12.2"
tqdm = "4.66.1"
pygithub = "^1.58.2"
mlflow = "2.8.1"
hydra-core = "^1.3.2"
numba = "0.57.1"
lmfit = "1.2.2"
pykalman = "0.9.5"
gluonts = "^0.13.4"
mxnet = "^1.9.1"
pytorch-forecasting = "0.10.3"
lightning = "^2.0.8"
urllib3 = "<2.0"
termcolor = "^2.3.0"
tensorly = "^0.8.1"
tensorboard = "^2.14.0"
pillow = "9.5.0"
dill = "0.3.5"
imageio = "2.33.1"
reformer-pytorch = "^1.4.4"
einops = "^0.7.0"
lingam = "^1.8.2"
graphviz = "0.8.4"
igraph = "^0.11.5"
cairocffi = "^1.7.1"
torch = {version = "<2.0", source = "torch_cu117"}
torchvision = {version = "<0.17.0", source = "torch_cu117"}
torchaudio = {version = "<2.2.0", source = "torch_cu117"}
cdt = "^0.6.0"
scipy = "1.11.4"
notears = {git = "https://github.com/xunzheng/notears.git"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
taskipy = "^1.10.1"
mypy = "^0.990"
pep8-naming = "^0.12.1"
pytest = "^7.1.1"
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"
Sphinx = "^4.5.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-pyproject = "^0.1.0"

[tool.ruff]
target-version = "py39"
line-length = 99
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "W", "F", "C"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = false
warn_unused_ignores = true
warn_redundant_casts = true

[tool.sphinx-pyproject]
project = "ModePlait"
copyright = "2023, naoki"
language = "en"
package_root = "ModePlait"
html_theme = "sphinx_rtd_theme"
todo_include_todos = true
templates_path = ["_templates"]
html_static_path = ["_static"]
extensions = [
  "sphinx.ext.autodoc",
  "sphinx.ext.viewcode",
  "sphinx.ext.todo",
  "sphinx.ext.napoleon",
]
